Java code/code/example-bicycle-lesson-6/example-bicycle-lesson-6/example-bicycle/src/main/java/fact/it/examplebicycle/model/Reservation.java
package fact.it.examplebicycle.model;
import java.time.LocalDate;

public class Reservation {
    private Bicycle bicycle;
    private LocalDate startDate;

    public Reservation() {
    }

    public Bicycle getBicycle() {
        return bicycle;
    }

    public void setBicycle(Bicycle bicycle) {
        this.bicycle = bicycle;
    }

    public LocalDate getStartDate() {
        return startDate;
    }

    public void setStartDate(LocalDate startDate) {
        this.startDate = startDate;
    }

    public String getDate() {
        return startDate.getDayOfMonth() + "/" + startDate.getMonthValue() + "/" + startDate.getYear();
    }
}
